remove event-sending logic to QueueService
